@@ -1,4 +1,4 @@
-class AddPosToEvents < ActiveRecord::Migration
+class AddPosToEvents < ActiveRecord::Migration[4.2]
   def change
     add_column :events, :position, :integer, default: 0, null: false
   end
