@@ -1,4 +1,4 @@
-class AddIndexToEventsOnPosAndParentId < ActiveRecord::Migration
+class AddIndexToEventsOnPosAndParentId < ActiveRecord::Migration[4.2]
   def change
     add_index :events, [:parent_id, :position], order: :position, where: "(parent_id IS NOT NULL)"
   end
