@@ -1,4 +1,4 @@
-class AddVersionToUsageReports < ActiveRecord::Migration
+class AddVersionToUsageReports < ActiveRecord::Migration[4.2]
   def change
     add_column :usage_reports, :version, :string
   end
