@@ -1,4 +1,4 @@
-class AddAuthorToDocument < ActiveRecord::Migration
+class AddAuthorToDocument < ActiveRecord::Migration[4.2]
   def change
     add_column :documents, :author_id, :integer, null: false, index: true
   end
