@@ -1,4 +1,4 @@
-class MoveDiscussionResources < ActiveRecord::Migration
+class MoveDiscussionResources < ActiveRecord::Migration[4.2]
   def change
     add_column :attachments, :migrated_to_document, :boolean, default: false, null: false
   end
